PB/requirements.txt
django==4.1.3
Pillow==9.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-dateutil==2.8.2